+++ microcosm_postgres/types/__init__.py
@@ -1,2 +1,5 @@
-from microcosm_postgres.types.sqlalchemy_utils_types import JSONType, UUIDType  # noqa: F401
-from microcosm_postgres.types.types import EnumType, Serial  # noqa: F401
+from microcosm_postgres.types.enum import EnumType  # noqa: F401
+from microcosm_postgres.types.json import JSONType  # noqa: F401
+from microcosm_postgres.types.serial import Serial  # noqa: F401
+from microcosm_postgres.types.utc_date_time import EPOCH, UTCDateTime, utcnow  # noqa: F401
+from microcosm_postgres.types.uuid import UUIDType  # noqa: F401
