+++ microcosm_postgres/types/json.py
@@ -1,0 +1,62 @@
+"""
+Copyright (c) 2012, Konsta Vesterinen
+
+Code from SQLAlchemy-Utils moved here until the library will support SQLAlchemy >= 1.4
+Related github issue: https://github.com/kvesteri/sqlalchemy-utils/issues/505
+
+"""
+import anyjson as json
+import sqlalchemy as sa
+from sqlalchemy.dialects.postgresql import JSON
+
+
+class JSONType(sa.types.TypeDecorator):
+    """
+    JSONType offers way of saving JSON data structures to database. On
+    PostgreSQL the underlying implementation of this data type is 'json' while
+    on other databases its simply 'text'.
+
+    Usage:
+
+        from sqlalchemy_utils import JSONType
+
+
+        class Product(Base):
+            __tablename__ = 'product'
+            id = sa.Column(sa.Integer, autoincrement=True)
+            name = sa.Column(sa.Unicode(50))
+            details = sa.Column(JSONType)
+
+
+        product = Product()
+        product.details = {
+            'color': 'red',
+            'type': 'car',
+            'max-speed': '400 mph'
+        }
+        session.commit()
+
+    """
+    impl = sa.UnicodeText
+
+    def load_dialect_impl(self, dialect):
+        if dialect.name == 'postgresql':
+            # Use the native JSON type.
+            return dialect.type_descriptor(JSON())
+
+        else:
+            return dialect.type_descriptor(self.impl)
+
+    def process_bind_param(self, value, dialect):
+        if dialect.name == 'postgresql':
+            return value
+        if value is not None:
+            value = json.dumps(value)
+        return value
+
+    def process_result_value(self, value, dialect):
+        if dialect.name == 'postgresql':
+            return value
+        if value is not None:
+            value = json.loads(value)
+        return value
